Battle Ship: smart AI attack
Clear AI attack queue and revert to random strike
when the biggest human ship is eliminated.

Still have to find a way to emulate human behavior
when ships of equal sizes are destroyed in an order different
than creation order. In such scenario the computer continues
to hit adjacent targets while a human would not do that.